+++ R/write.R
@@ -2,6 +2,7 @@
     x,
     dataset_dir = tempdir(),
     dataset_name = "object.yaml",
+    overwrite = TRUE,
     verbose = TRUE,
     ...
 ) {
@@ -13,14 +14,25 @@
         paste0(".yaml") |>
         normalizePath(mustWork = FALSE)
 
+    if (file.exists(output_path) && !overwrite) {
+        if (verbose) {
+            message(
+                "The file already exists and won't be overwritten.",
+                "To overwrite this file, please set `overwrite = TRUE`."
+            )
+        }
+        return(invisible(FALSE))
+    }
+
+    if (file.exists(output_path) && verbose) {
+        message("The file already exists and will be overwritten.")
+    }
+
     if (!dir.exists(dataset_dir)) {
         dir.create(dataset_dir)
     }
     if (verbose) {
         message("The datasets will be exported to ", output_path, ".")
-        if (file.exists(output_path)) {
-            message("The file already exists and will be overwritten.")
-        }
     }
     x_yaml <- yaml::as.yaml(x, precision = 22L, indent = 2L)
     writeLines(
@@ -28,7 +40,7 @@
         con = output_path,
         useBytes = TRUE
     )
-    return(invisible(TRUE))
+    return(invisible(output_path))
 }
 
 #' @title Save datasets in a yaml file
@@ -36,6 +48,8 @@
 #' @param x an object of class \code{IssuesTB}, \code{LabelsTB} or
 #' \code{MilestonesTB}.
 #' @inheritParams get_issues
+#' @param overwrite Boolean. If the dataset file already exists,
+#'   should it be overwrite? Default is TRUE.
 #' @param \dots Unused parameter.
 #'
 #' @details
@@ -83,7 +97,14 @@
 #'
 #' @rdname write
 #'
-write_to_dataset <- function(x, ...) {
+write_to_dataset <- function(
+    x,
+    dataset_dir = getOption("IssueTrackeR.dataset.dir"),
+    dataset_name,
+    overwrite = TRUE,
+    verbose = TRUE,
+    ...
+) {
     UseMethod(generic = "write_to_dataset", object = x)
 }
 
@@ -95,10 +116,11 @@ write_to_dataset.IssuesTB <- function(
     x,
     dataset_dir = getOption("IssueTrackeR.dataset.dir"),
     dataset_name = "list_issues.yaml",
+    overwrite = TRUE,
     verbose = TRUE,
     ...
 ) {
-    .write(x, dataset_dir, dataset_name, verbose)
+    .write(x, dataset_dir, dataset_name, overwrite, verbose)
     return(invisible(TRUE))
 }
 
@@ -111,10 +133,11 @@ write_to_dataset.LabelsTB <- function(
     x,
     dataset_dir = getOption("IssueTrackeR.dataset.dir"),
     dataset_name = "list_labels.yaml",
+    overwrite = TRUE,
     verbose = TRUE,
     ...
 ) {
-    .write(x, dataset_dir, dataset_name, verbose)
+    .write(x, dataset_dir, dataset_name, overwrite, verbose)
     return(invisible(TRUE))
 }
 
@@ -126,10 +149,11 @@ write_to_dataset.MilestonesTB <- function(
     x,
     dataset_dir = getOption("IssueTrackeR.dataset.dir"),
     dataset_name = "list_milestones.yaml",
+    overwrite = TRUE,
     verbose = TRUE,
     ...
 ) {
-    .write(x, dataset_dir, dataset_name, verbose)
+    .write(x, dataset_dir, dataset_name, overwrite, verbose)
     return(invisible(TRUE))
 }
 
@@ -137,7 +161,7 @@ write_to_dataset.MilestonesTB <- function(
 #' @exportS3Method write_to_dataset default
 #' @method write_to_dataset default
 #' @export
-write_to_dataset.default <- function(x, ...) {
+write_to_dataset.default <- function(...) {
     stop(
         "This function requires a IssuesTB, LabelsTB or MilestonesTB object.",
         call. = FALSE