+++ NEWS.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+* `overwrite` argument to specify if we want to overwrite an existing file if it already exists. (#18)
+
 
 ## [1.4.0] - 2026-06-21